examples/agents/python/google-adk/requirements.txt
google-adk==1.19.0
